requirements.txt
asgiref==3.10.0
Django==5.2.7
djangorestframework==3.16.1
django-filter==24.3
psycopg==3.2.11
psycopg-binary==3.2.11
python-dotenv==1.1.1
sqlparse==0.5.3
typing_extensions==4.15.0
